feat: modify endpoints so that you can't update a published campaign

+++ backend/server/src/handler/campaign.rs
@@ -115,6 +115,12 @@ impl CampaignHandler {
         _admin: CampaignAdmin,
         Json(request_body): Json<models::campaign::CampaignUpdate>,
     ) -> Result<impl IntoResponse, ChaosError> {
+        let campaign = Campaign::get(attachment.campaign_id, &mut transaction.tx).await?;
+
+        if campaign.published {
+            return Err(ChaosError::BadRequest);
+        }
+
         Campaign::update(id, request_body, &mut transaction.tx).await?;
         transaction.tx.commit().await?;
         Ok(AppMessage::OkMessage("Successfully updated campaign"))
@@ -138,6 +144,11 @@ impl CampaignHandler {
         Path(id): Path<i64>,
         _admin: CampaignAdmin,
     ) -> Result<impl IntoResponse, ChaosError> {
+        let campaign = Campaign::get(attachment.campaign_id, &mut transaction.tx).await?;
+
+        if campaign.published {
+            return Err(ChaosError::BadRequest);
+        }
         Campaign::publish(id, &mut transaction.tx).await?;
         transaction.tx.commit().await?;
         Ok(AppMessage::OkMessage("Successfully published campaign"))
@@ -163,6 +174,12 @@ impl CampaignHandler {
         Path(id): Path<i64>,
         _admin: CampaignAdmin,
     ) -> Result<impl IntoResponse, ChaosError> {
+        let campaign = Campaign::get(attachment.campaign_id, &mut transaction.tx).await?;
+
+        if campaign.published {
+            return Err(ChaosError::BadRequest);
+        }
+
         let banner_url =
             Campaign::update_banner(id, &mut transaction.tx, &state.storage_bucket).await?;
         transaction.tx.commit().await?;
@@ -187,6 +204,12 @@ impl CampaignHandler {
         Path(id): Path<i64>,
         _admin: CampaignAdmin,
     ) -> Result<impl IntoResponse, ChaosError> {
+        let campaign = Campaign::get(attachment.campaign_id, &mut transaction.tx).await?;
+
+        if campaign.published {
+            return Err(ChaosError::BadRequest);
+        }
+
         Campaign::delete(id, &mut transaction.tx).await?;
         transaction.tx.commit().await?;
         Ok(AppMessage::OkMessage("Successfully deleted campaign"))
@@ -214,6 +237,12 @@ impl CampaignHandler {
         _admin: CampaignAdmin,
         Json(data): Json<RoleUpdate>,
     ) -> Result<impl IntoResponse, ChaosError> {
+        let campaign = Campaign::get(attachment.campaign_id, &mut transaction.tx).await?;
+
+        if campaign.published {
+            return Err(ChaosError::BadRequest);
+        }
+
         Role::create(id, data, &mut transaction.tx, &mut state.snowflake_generator).await?;
         transaction.tx.commit().await?;
         Ok(AppMessage::OkMessage("Successfully created role"))
@@ -431,6 +460,12 @@ impl CampaignHandler {
         _admin: CampaignAdmin,
         Json(data): Json<Vec<NewAttachment>>,
     ) -> Result<impl IntoResponse, ChaosError> {
+        let campaign = Campaign::get(attachment.campaign_id, &mut transaction.tx).await?;
+
+        if campaign.published {
+            return Err(ChaosError::BadRequest);
+        }
+
         let upload_results = CampaignAttachment::create_or_update_multiple(id, data, &mut transaction.tx, &mut state.snowflake_generator, &state.storage_bucket).await?;
         transaction.tx.commit().await?;
         Ok((StatusCode::OK, Json(upload_results)))
@@ -472,7 +507,11 @@ impl CampaignHandler {
         // Get the attachment to find its campaign_id
         let attachment = CampaignAttachment::get_by_id(attachment_id, &mut transaction.tx).await?;
         let campaign = Campaign::get(attachment.campaign_id, &mut transaction.tx).await?;
-        
+
+        if campaign.published {
+            return Err(ChaosError::BadRequest);
+        }
+
         // Verify the admin has access to this campaign
         crate::service::campaign::user_is_campaign_admin(
             user.user_id,
